Event System Implemented
Events are implemented so when another script listens, they're able to know when the stretching began, when the stretching ended, when the shrinking began, and when the shrinking ended. An example of this in action is when the shrink begins another event is invoked telling the player to jump which makes the stretch mechanic fun and more functional.